Move auth mock data into mock_data.py
